Update build to C++11 (gnu++11) with "-O3".
Change build flags from "-Ofast" (Level 4) to "-O3" (Level 3) using
C++11 mode with GNU extensions (gnu++11), printing the clang command.
To fix new build errors, spaces were added to do assertion macros in
test/task.h and an additional cast was added in src/unix/vms-poll.c.

+++ src/unix/vms-poll.c
@@ -403,7 +403,7 @@ void uv__io_poll(uv_loop_t* loop, int timeout) {
       }
     } else {
       /* Set timer ef to wake on timeout, then sleep on either event flag. */
-      struct _generic_64 delta_timeout = { -(timeout * 10000) };
+      struct _generic_64 delta_timeout = { (uint64_t) (-((int64_t) timeout * 10000)) };
       if (!$VMS_STATUS_SUCCESS(sys$setimr(loop->timer_ef, &delta_timeout,
                         NULL, (int64_t) loop->timer_ef, 0))) {
         abort();
